FIX - Process filter on log page
